Move (content-blocks) : Folder Segmentation
