+++ app/components/contents/columns/ColumnList.tsx
@@ -1,13 +1,13 @@
-import { ColumnBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
+import { ColumnListBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
 
 import { fetchBlocks } from "@/app/lib/databases";
-import Block from "./Block";
+import Block from "../Block";
 
-export default async function Column({ block }: { block: ColumnBlockObjectResponse }) {
+export default async function ColumnList({ block }: { block: ColumnListBlockObjectResponse }) {
     const children = block.has_children ? await fetchBlocks(block.id) : null;
 
     return (
-        <div className="article-content--column">
+        <div className="article-content--column-list">
             {children?.items.map((item, index) =>
                 <Block key={index} block={item} blocks={children?.items} index={index} />)}
         </div>
